[183046] Should not require to build axis2.war and other bugs.

+++ axis2/bundles/org.eclipse.jst.ws.axis2.core/src/org/eclipse/jst/ws/axis2/facet/commands/Axis2WebservicesServerCommand.java
@@ -11,6 +11,7 @@
  * -------- -------- -----------------------------------------------------------
  * 20070213  168766 sandakith@wso2.com - Lahiru Sandakith, Initial code to introduse the Axis2 
  * 										  facet to the framework for 168766
+ * 20070426   183046 sandakith@wso2.com - Lahiru Sandakith
  *******************************************************************************/
 package org.eclipse.jst.ws.axis2.facet.commands;
 
@@ -21,10 +22,12 @@ import org.eclipse.core.runtime.IAdaptable;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
+import org.eclipse.jst.ws.axis2.core.plugin.data.ServerModel;
 import org.eclipse.jst.ws.axis2.core.plugin.messages.Axis2CoreUIMessages;
 import org.eclipse.jst.ws.axis2.core.utils.FacetContainerUtils;
 import org.eclipse.jst.ws.axis2.core.utils.FileUtils;
 import org.eclipse.jst.ws.axis2.facet.utils.Axis2RuntimeUtils;
+import org.eclipse.jst.ws.axis2.facet.utils.Axis2WebappUtils;
 import org.eclipse.jst.ws.axis2.facet.utils.ContentCopyUtils;
 import org.eclipse.wst.common.frameworks.datamodel.AbstractDataModelOperation;
 
@@ -44,11 +47,18 @@ AbstractDataModelOperation {
 	public IStatus exexuteOverrride(IProgressMonitor monitor) throws ExecutionException {
 		//Copy the axis2 libs in to this client project
 		IStatus status = Status.OK_STATUS;
-		
+		String runtimeLocation = null;
 		ContentCopyUtils contentCopyUtils = new ContentCopyUtils();
-				String runtimeLocation = Axis2RuntimeUtils.copyAxis2War(
+		
+		if(ServerModel.isAxis2ServerPathRepresentsWar()){
+			runtimeLocation = Axis2RuntimeUtils.copyAxis2War(
 															monitor,
 															Axis2CoreUIMessages.PROPERTY_KEY_PATH);
+		}else{
+			runtimeLocation = Axis2WebappUtils.copyAxis2War(
+					monitor,
+					Axis2CoreUIMessages.PROPERTY_KEY_PATH);
+		}
 				status = contentCopyUtils.copyDirectoryRecursivelyIntoWorkspace(
 						runtimeLocation, 
 						FacetContainerUtils.pathToWebProjectContainer(project), 
@@ -57,7 +67,7 @@ AbstractDataModelOperation {
 				
 				//clean up tempory files
 				File tempFacetDirectory = new File(runtimeLocation);
-				if (tempFacetDirectory.exists()) {
+				if (tempFacetDirectory.exists() && ServerModel.isAxis2ServerPathRepresentsWar()) {
 					FileUtils.deleteDir(tempFacetDirectory);
 				}
 		return status;
